setMainVersion: don't update if version number already ok (avoid touching file unnecessarily)
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@10381 a380766d-ff14-0410-b294-a243070f3f08

+++ common/setMainVersion.sh
@@ -16,7 +16,9 @@ line=$( grep "classvar scVersion" $mainfile )
 echo $versionline
 echo $line
 
-sed "s/$line/$versionline/" <$mainfile >tempfile
-cp tempfile $mainfile
-rm tempfile
-
+# only update if they don't already match
+if [ "$versionline" != "$line" ]; then
+	sed "s/$line/$versionline/" <$mainfile >tempfile
+	cp tempfile $mainfile
+	rm tempfile
+fi
